1.保存抢鞋用户时先判断是否存在，存在则返回失败。 2.fix job的结束逻辑

+++ dao/src/main/java/com/djcao/boot/repository/ShoesItemRepository.java
@@ -23,5 +23,5 @@ public interface ShoesItemRepository extends JpaRepository<ShoesItem,Long> {
     int updateStatusByItemIdList(@Param("itemIdList")List<String> itemIdList, @Param("status")Integer status);
 
     @Query(value = "select t from ShoesItem t where t.status = :status")
-    Page<ShoesItem> findByStatus(@Param("status") Integer status,Pageable pageable);
+    Page<ShoesItem> findByStatus(@Param("status") Byte status,Pageable pageable);
 }
